app/requirements.txt
requests
python-dotenv
openai
semantic-kernel